requirements.txt
streamlit
torch
torchvision
opencv-python-headless
Pillow
numpy
matplotlib
gdown